requirements.txt
torch
torchvision
Pillow
numpy
scikit-learn
opencv-python
ultralytics
git+https://github.com/openai/CLIP.git